LeaveCredits

+++ ProcessLayer/Entities/Lookups/LeaveDefaultCredits.cs
@@ -4,8 +4,8 @@
     {
         public int ID { get; set; }
         public byte LeaveTypeID { get; set; }
-        public byte MinYearsInService { get; set; }
-        public byte? MaxYearsInService { get; set; }
+        public float MinYearsInService { get; set; }
+        public float? MaxYearsInService { get; set; }
         public float Credits { get; set; }
         public LeaveType LeaveType { get; set; }
     }
